+++ packages/runtime/container-runtime/src/summary/summaryDelayLoadedModule/summaryGenerator.ts
@@ -3,205 +3,39 @@
  * Licensed under the MIT License.
  */
 
-import { ITelemetryBaseProperties } from "@fluidframework/core-interfaces";
-import {
-	assert,
-	Deferred,
-	IPromiseTimer,
-	IPromiseTimerResult,
-	Timer,
-} from "@fluidframework/core-utils/internal";
+import { assert, IPromiseTimer, Timer } from "@fluidframework/core-utils/internal";
 import { DriverErrorTypes, MessageType } from "@fluidframework/driver-definitions/internal";
 import { getRetryDelaySecondsFromError } from "@fluidframework/driver-utils/internal";
 import {
 	isFluidError,
 	ITelemetryLoggerExt,
-	LoggingError,
 	PerformanceEvent,
 	wrapError,
 } from "@fluidframework/telemetry-utils/internal";
 
-import {
-	IAckSummaryResult,
-	IBroadcastSummaryResult,
-	INackSummaryResult,
+import type {
 	IRefreshSummaryAckOptions,
+	IRetriableFailureError,
 	ISubmitSummaryOptions,
 	ISummarizeHeuristicData,
-	ISummarizeResults,
-	ISummaryCancellationToken,
 	SubmitSummaryFailureData,
 	SubmitSummaryResult,
-	SummarizeResultPart,
 	SummaryGeneratorTelemetry,
-	type IRetriableFailureError,
-} from "./summarizerTypes.js";
-import { IClientSummaryWatcher } from "./summaryCollection.js";
+} from "../summarizerTypes.js";
+import {
+	RetriableSummaryError,
+	getFailMessage,
+	raceTimer,
+	type SummarizeErrorCode,
+} from "../summarizerUtils.js";
+import type { IClientSummaryWatcher } from "../summaryCollection.js";
 
-export type raceTimerResult<T> =
-	| { result: "done"; value: T }
-	| { result: IPromiseTimerResult["timerResult"] }
-	| { result: "cancelled" };
-
-/**
- * Wait for a promise or PromiseTimer to elapse.
- */
-export async function raceTimer<T>(
-	promise: Promise<T>,
-	timer: Promise<IPromiseTimerResult>,
-
-	cancellationToken?: ISummaryCancellationToken,
-): Promise<raceTimerResult<T>> {
-	const promises: Promise<raceTimerResult<T>>[] = [
-		promise.then((value) => ({ result: "done", value }) as const),
-		timer.then(({ timerResult: result }) => ({ result }) as const),
-	];
-	if (cancellationToken !== undefined) {
-		promises.push(
-			cancellationToken.waitCancelled.then(() => ({ result: "cancelled" }) as const),
-		);
-	}
-	return Promise.race(promises);
-}
+import { SummarizeResultBuilder } from "./summaryResultBuilder.js";
+import { type INackSummaryResult, type ISummarizeResults } from "./summaryResultTypes.js";
 
 // Send some telemetry if generate summary takes too long
 const maxSummarizeTimeoutTime = 20000; // 20 sec
 const maxSummarizeTimeoutCount = 5; // Double and resend 5 times
-
-export type SummarizeReason =
-	/**
-	 * Attempt to summarize after idle timeout has elapsed.
-	 * Idle timer restarts whenever an op is received. So this
-	 * triggers only after some amount of time has passed with
-	 * no ops being received.
-	 */
-	| "idle"
-	/**
-	 * Attempt to summarize after a maximum time since last
-	 * successful summary has passed. This measures time since
-	 * last summary ack op was processed.
-	 */
-	| "maxTime"
-	/**
-	 * Attempt to summarize after a maximum number of ops have
-	 * passed since the last successful summary. This compares
-	 * op sequence numbers with the reference sequence number
-	 * of the summarize op corresponding to the last summary
-	 * ack op.
-	 */
-	| "maxOps"
-	/**
-	 * Special case to attempt to summarize one last time before the
-	 * summarizer client closes itself. This is to prevent cases where
-	 * the summarizer client never gets a chance to summarize, because
-	 * there are too many outstanding ops and/or parent client cannot
-	 * stay connected long enough for summarizer client to catch up.
-	 */
-	| "lastSummary"
-	/**
-	 * On-demand summary requested with specified reason.
-	 */
-	| `onDemand;${string}`
-	/**
-	 * Enqueue summarize attempt with specified reason.
-	 */
-	| `enqueue;${string}`;
-
-const summarizeErrors = {
-	/**
-	 * Error encountered while generating the summary tree, uploading
-	 * it to storage, or submitting the op. It could be a result of
-	 * the client becoming disconnected while generating or an actual error.
-	 */
-	submitSummaryFailure: "Error while generating, uploading, or submitting summary",
-	/**
-	 * The summaryAckWaitTimeout time has elapsed before receiving the summarize op
-	 * sent by this summarize attempt. It is expected to be broadcast quickly.
-	 */
-	summaryOpWaitTimeout: "Timeout while waiting for summarize op broadcast",
-	/**
-	 * The summaryAckWaitTimeout time has elapsed before receiving either a
-	 * summaryAck or summaryNack op from the server in response to this
-	 * summarize attempt. It is expected that the server should respond.
-	 */
-	summaryAckWaitTimeout: "Timeout while waiting for summaryAck/summaryNack op",
-	/**
-	 * The server responded with a summaryNack op, thus rejecting this
-	 * summarize attempt.
-	 */
-	summaryNack: "Server rejected summary via summaryNack op",
-
-	disconnect: "Summary cancelled due to summarizer or main client disconnect",
-} as const;
-
-export type SummarizeErrorCode = keyof typeof summarizeErrors;
-
-// Helper functions to report failures and return.
-export const getFailMessage = (errorCode: SummarizeErrorCode): string =>
-	`${errorCode}: ${summarizeErrors[errorCode]}`;
-
-export class SummarizeResultBuilder {
-	public readonly summarySubmitted = new Deferred<
-		SummarizeResultPart<SubmitSummaryResult, SubmitSummaryFailureData>
-	>();
-	public readonly summaryOpBroadcasted = new Deferred<
-		SummarizeResultPart<IBroadcastSummaryResult>
-	>();
-	public readonly receivedSummaryAckOrNack = new Deferred<
-		SummarizeResultPart<IAckSummaryResult, INackSummaryResult>
-	>();
-
-	/**
-	 * Fails one or more of the three results as per the passed params.
-	 * If submit fails, all three results fail.
-	 * If op broadcast fails, only op broadcast result and ack nack result fails.
-	 * If ack nack fails, only ack nack result fails.
-	 */
-	public fail(
-		message: string,
-		error: IRetriableFailureError,
-		submitFailureResult?: SubmitSummaryFailureData,
-		nackSummaryResult?: INackSummaryResult,
-	): void {
-		assert(
-			!this.receivedSummaryAckOrNack.isCompleted,
-			0x25e /* "no reason to call fail if all promises have been completed" */,
-		);
-
-		const result: SummarizeResultPart<undefined> = {
-			success: false,
-			message,
-			data: undefined,
-			error,
-		} as const;
-
-		// Note that if any of these are already resolved, it will be a no-op. For example, if ack nack failed but
-		// submit summary and op broadcast has already been resolved as passed, only ack nack result will get modified.
-		this.summarySubmitted.resolve({ ...result, data: submitFailureResult });
-		this.summaryOpBroadcasted.resolve(result);
-		this.receivedSummaryAckOrNack.resolve({ ...result, data: nackSummaryResult });
-	}
-	public build(): ISummarizeResults {
-		return {
-			summarySubmitted: this.summarySubmitted.promise,
-			summaryOpBroadcasted: this.summaryOpBroadcasted.promise,
-			receivedSummaryAckOrNack: this.receivedSummaryAckOrNack.promise,
-		} as const;
-	}
-}
-
-/**
- * Errors type for errors hit during summary that may be retriable.
- */
-export class RetriableSummaryError extends LoggingError implements IRetriableFailureError {
-	constructor(
-		message: string,
-		public readonly retryAfterSeconds?: number,
-		props?: ITelemetryBaseProperties,
-	) {
-		super(message, props);
-	}
-}
 
 /**
  * This class generates and tracks a summary attempt.
